Cleaning up code for SvarFci and SvarGfci.
